+++ src/ansys/health/heart/writer/_control_volume.py
@@ -23,8 +23,8 @@
 
 from dataclasses import dataclass
 
-from ansys.health.heart.anatomical_parts import _Chamber
 from ansys.health.heart.models import BiVentricle, FourChamber, HeartModel, LeftVentricle
+from ansys.health.heart.parts import _Chamber
 from ansys.health.heart.writer.define_function_templates import _define_function_0d_system
